requirements.txt
streamlit>=1.30.0
openai>=1.3.0
playwright>=1.40.0
python-dotenv
pymupdf  # fitz olarak import edilir
fpdf2
pdf2image
requests